+++ src/core.h
@@ -71,6 +71,19 @@ void raiseSystemError(DSState &st, int errorNum, std::string &&message);
 void raiseShellExit(DSState &st, int64_t status);
 
 /**
+ * print error message with current location (source:lineno)
+ * emit newline
+ * @param state
+ * @param errNum
+ * may be 0
+ * @param fmt
+ * @param ...
+ * @return
+ */
+bool printErrorAt(const DSState &state, int errNum, const char *fmt, ...)
+    __attribute__((format(printf, 3, 4)));
+
+/**
  * get and set signal handler.
  * if handler is null, not set handler.
  * @param st